pyproject.toml
[project]
name = "bedrock-agent-a2a"
version = "0.1.0"
description = "Simple AWS Bedrock Agent with ADK exposed via A2A protocol"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "boto3>=1.35.0",
    "fasta2a>=0.1.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.1",
    "redis>=7.0.1",
    "google-adk>=1.18.0",
    "sqlmodel>=0.0.16",
    "typer>=0.12.0",
    "fastmcp>=2.13.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
